lab/lab2/第二次作业_nlp部分/attacks_nlp/requirements.txt
python 3.7  
pytorch 1.1  
tqdm  
sklearn  
tensorboardX